Update Unit 3_Pir sensor
instead of printing the motion messages to the serial console, I print the raw reading of the sensor. Please use the Tools > Serial Graph on your Arduino programming environment (IDE) to visualize how the sensor behaves across time. That should give you an idea is the delay problem is in the code or in the sensor.